Move mint data to raw transaction class
