+++ ironfish/src/rpc/routes/wallet/createTransaction.ts
@@ -5,8 +5,7 @@ import { Asset } from '@ironfish/rust-nodejs'
 import { BufferMap } from 'buffer-map'
 import * as yup from 'yup'
 import { BurnDescription } from '../../../primitives/burnDescription'
-import { MintData } from '../../../primitives/mintData'
-import { RawTransactionSerde } from '../../../primitives/rawTransaction'
+import { MintData, RawTransactionSerde } from '../../../primitives/rawTransaction'
 import { CurrencyUtils } from '../../../utils'
 import { NotEnoughFundsError } from '../../../wallet/errors'
 import { ERROR_CODES, ValidationError } from '../../adapters/errors'
