+++ ironfish/src/testUtilities/fixtures/transactions.ts
@@ -5,7 +5,7 @@ import { Asset } from '@ironfish/rust-nodejs'
 import { Assert } from '../../assert'
 import { IronfishNode } from '../../node'
 import { BurnDescription } from '../../primitives/burnDescription'
-import { MintData } from '../../primitives/mintData'
+import { MintData } from '../../primitives/rawTransaction'
 import { SerializedTransaction, Transaction } from '../../primitives/transaction'
 import { Account, Wallet } from '../../wallet'
 import { createRawTransaction } from '../helpers/transaction'
