+++ ironfish/src/wallet/wallet.ts
@@ -25,10 +25,9 @@ import { NoteWitness, Witness } from '../merkletree/witness'
 import { Mutex } from '../mutex'
 import { BlockHeader } from '../primitives/blockheader'
 import { BurnDescription } from '../primitives/burnDescription'
-import { MintData } from '../primitives/mintData'
 import { Note } from '../primitives/note'
 import { NOTE_ENCRYPTED_SERIALIZED_SIZE_IN_BYTE } from '../primitives/noteEncrypted'
-import { RawTransaction } from '../primitives/rawTransaction'
+import { MintData, RawTransaction } from '../primitives/rawTransaction'
 import { SPEND_SERIALIZED_SIZE_IN_BYTE } from '../primitives/spend'
 import { Transaction } from '../primitives/transaction'
 import { IDatabaseTransaction } from '../storage/database/transaction'
